predictor/pyproject.toml
[project]
name = "simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.0",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "wandb>=0.19.1",
]
